1. Fixed logout causing untimely close of streamcontrollers 2. Implemented delete all schedules in mobile app 3. Implemented delete all schedules in device server

+++ petfeed/lib/src/bloc/schedules_bloc/schedules_bloc.dart
@@ -74,6 +74,11 @@ class SchedulesBloc extends Bloc {
     }
   }
 
+  void applyRecommendedClosed() {
+    preferences.setBool('close', true);
+    dispatch(ApplyRecommendedClosed());
+  }
+
   void applyRecommended({
     @required List<String> feedTimes,
     @required double amount,
@@ -129,13 +134,18 @@ class SchedulesBloc extends Bloc {
   }
 
   void deleteAllSchedules() {
-    dispatch(DeleteAllSchedules());
+    String deviceToken = preferences.getString('deviceToken');
+    dispatch(DeleteAllSchedules((b) => b..accessToken = deviceToken));
   }
 
   void _mapDeleteAllSchedules(DeleteAllSchedules event) async {
     try {
-      await initDB();
-      await provider.deleteAll();
+      final response =
+          await dataSource.deleteSchedules(accessToken: event.accessToken);
+      if (response.status == 'success') {
+        await initDB();
+        await provider.deleteAll();
+      }
       getSchedules();
     } catch (_) {
       print(_);
